Add admin mixin
